+++ src/features/family-management/model/types.ts
@@ -1,0 +1,5 @@
+import type { CreatePetInvitationCodeResponse } from '@/features/auth';
+
+export interface InvitationCodeState extends CreatePetInvitationCodeResponse {
+  createdAt: number;
+}
